# Memberlist: Optimise receive path by not cloning current ring state.