--- a/kv/memberlist/mergeable.go
+++ b/kv/memberlist/mergeable.go
@@ -9,6 +9,9 @@ type Mergeable interface {
 	// If merge doesn't result in any change, returns nil.
 	// Error can be returned if merging with given 'other' value is not possible.
 	//
+	// Implementations may modify the internal state even if a change is not returned,
+	// so long as the logical value does not change, i.e. the properties below are maintained.
+	//
 	// In order for state merging to work correctly, Merge function must have some properties. When talking about the
 	// result of the merge in the following text, we don't mean the return value ("change"), but the
 	// end-state of receiver. That means Result of A.Merge(B) is end-state of A.