**What this PR does**:

When memberlist is used as the backing store for the ring, processing updates
can be costly if the ring state is very large. This is partly due to the ring
state being cloned upon the arrival of every message. It is believed that this
can be avoided. Additionally, the reduced memory allocations will reduce the
amount of work required by the GC.

The included benchmark shows a good improvement:

Command:
```
dskit/bench$ go test -v -bench BenchmarkMemberlistReceiveWithRingDesc -benchmem -cpu 6
```

```
name                             old time/op    new time/op    delta
MemberlistReceiveWithRingDesc-6    1.76ms ± 5%    1.01ms ± 6%  -42.32%  (p=0.000 n=10+9)

name                             old alloc/op   new alloc/op   delta
MemberlistReceiveWithRingDesc-6     566kB ± 0%     254kB ± 0%  -55.09%  (p=0.000 n=9+10)

name                             old allocs/op  new allocs/op  delta
MemberlistReceiveWithRingDesc-6     1.86k ± 0%     0.64k ± 0%  -65.83%  (p=0.000 n=10+10)
```

_(Edit: Updated results to take into account adding tokens to benchmark)_

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->


**Which issue(s) this PR fixes**:

Fixes N/A

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`