@pstibrany 2021-11-23
I'm getting race conditions when running memberlist tests with `-race`.