@stevesg 2021-11-23
I've added a fix to Clone the `changes` while holding the lock before returning them. I think this is necessary because the `Mergeable` implementors are allowed to reference the source state in the returned changes. For example, in `ring.Desc`, the `Tokens` slice can be shared and is not deep copied. For safety, we can just clone the whole thing, given the changes are typically small.

The performance appears to be unchanged:

```
BenchmarkMemberlistReceiveWithRingDesc-6            3229            324260 ns/op          239825 B/op         37 allocs/op
```